chore: fixed iimport/export path for the moved Product components
